Fix: replies are too slow
We only send a reply if we have not replied before. We do this by
searching Twitter for replies. However, because this is unreliable
because it depends on Twitter updating their search indexes. To avoid
this we added a 30 minute delay between updates.

This change makes reply-guy faster by also storing a record of tweets
we have replied to, so we can check more regularly.